requirements.txt
# Core
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.15
alembic==1.11.1
psycopg2-binary==2.9.6

# File handling
python-magic==0.4.27
python-magic-bin==0.4.14; sys_platform == 'win32'

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
pydantic[email]==1.10.7

# Testing
pytest==7.3.1
pytest-cov==4.0.0
httpx==0.24.0

# Development
black==23.3.0
isort==5.12.0
flake8==6.0.0
mypy==1.2.0

# Logging
python-json-logger==2.0.7

# Async
anyio==3.6.2

# Social Media APIs
facebook-sdk==3.1.0
python-instagram==1.3.2

# Image Processing
Pillow==9.5.0

# Rate Limiting
slowapi==0.1.8

# URL Validation
validators==0.20.0
